refactoring : interface를 abstract class로 변경
